fixed wrong return value

+++ ESP32/window_alert/src/hardware/WindowSensor.cpp
@@ -69,7 +69,7 @@ public:
         return this->mGpioOutput;
     }
 
-    int getTimestampLastInterrupt() {
+    unsigned long getTimestampLastInterrupt() {
         return this->mTimestampLastInterrupt;
     }
